coordinator/requirements.txt
flask
requests
numpy
pillow
gunicorn